chore(ott): clean up in extension
