+++ apps/extension/src/lib/components/sidebar-tab-context-menu.svelte
@@ -62,7 +62,7 @@
   </ContextMenu.Trigger>
   <ContextMenu.Portal>
     <ContextMenu.Content>
-      <ul class="menu menu-sm bg-base-200 rounded-box w-56">
+      <ul class="menu menu-sm bg-base-300 rounded-box w-56">
         <li>
           <a onclick={handleNewTab}>
             <PlusIcon size={16} />
